Revert "Implement drag-based document multi selection"
This reverts commit 8cd92a5a8ead6cd1cb44cc395676c13293873f0d.

As drag'n'drop seems to be in a pretty sorry state in chrome/chromium
(pretty much no matter how you try to implement it), enabling this
feature completely breaks (non-add-all) document selection for chrome
users.

We'll have to think of something else.
(This does not revert Clippy. Clippy can stay (although disabled) until
we can think of a better way to explain how selection works to users)