Move most of index.html into separate template files.
This pulls in infuser (base templating support), TraffiCop (template
caching), and koExternalTemplateEngine (simple buildsystem-less external
templates https://github.com/ifandelse/Knockout.js-External-Template-Engine).

This approach has the downside of requiring a separate request for each
template included into the page, but has the upside of being easy to use
and simple to maintain, as it doesn't require a build system for
'compiling' the templates.