Fix grey autosuggestion text getting cut off
typeahead.js constructs a second input element for this, and as the id
gets lost on that copy, setting the width via #-selector won't work.
The !important is necessary due to the stupidity of CSS precedence rules
in combination with our build chain (something in bootstrap would overwrite
the width).