transcriptmigration: Keep a list of recent lectures/examinants
The 5 most recent lectures and examinants are provided in a list and
added on click. This should speed up the migration process as there are
many transcripts that share examinants or lectures.

NB: I use `$('#transcript-migration-lectures').tagsinput('add', $data);`
instead of modifying `recentLectures` directly because tagsinput does
not always correctly update this way. Also, I get the nice "already
exists" animation for free.